move dataloaderadapters to dataloaders
